# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  dto.RoleDescription:
    properties:
      description:
        type: string
      profile_id:
        type: integer
    type: object
  model.Certification:
    properties:
      certification_id:
        type: integer
      date_issued:
        type: string
      issuer:
        type: string
      name:
        type: string
      profile_id:
        type: integer
    type: object
  model.Contact:
    properties:
      contact_id:
        type: integer
      email:
        type: string
      phone_number:
        type: string
      profile_id:
        type: integer
    type: object
  model.Education:
    properties:
      degree:
        type: string
      education_id:
        type: string
      end_date:
        type: string
      field:
        type: string
      institution:
        type: string
      profile_id:
        type: integer
      start_date:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  model.Experience:
    properties:
      company:
        type: string
      description:
        type: string
      end_date:
        type: string
      experience_id:
        type: string
      is_current:
        type: boolean
      profile_id:
        type: integer
      role:
        type: string
      start_date:
        type: string
    type: object
  model.Language:
    properties:
      language_id:
        type: string
      level:
        type: string
      name:
        type: string
      profile_id:
        type: integer
    type: object
  model.ProfileDTO:
    properties:
      active:
        type: boolean
      birth_date:
        type: string
      certifications:
        items:
          $ref: '#/definitions/model.Certification'
        type: array
      contacts:
        items:
          $ref: '#/definitions/model.Contact'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      educations:
        items:
          $ref: '#/definitions/model.Education'
        type: array
      email:
        type: string
      experiences:
        items:
          $ref: '#/definitions/model.Experience'
        type: array
      languages:
        items:
          $ref: '#/definitions/model.Language'
        type: array
      name:
        type: string
      phone_number:
        type: string
      profile_id:
        type: integer
      projects:
        items:
          $ref: '#/definitions/model.Project'
        type: array
      skills:
        items:
          $ref: '#/definitions/model.SkillDTO'
        type: array
      socialMedias:
        items:
          $ref: '#/definitions/model.SocialMedia'
        type: array
      updated_at:
        type: string
    type: object
  model.Project:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      link:
        type: string
      name:
        type: string
      profile_id:
        type: integer
      project_id:
        type: string
      updated_at:
        type: string
    type: object
...
  model.SkillDTO:
    properties:
      embeddings:
        items:
          type: number
        type: array
      level:
        type: string
      name:
        type: string
      profile_id:
        type: integer
      skill_id:
        type: string
    type: object
  model.SocialMedia:
    properties:
      handle:
        type: string
      link:
        type: string
      platform:
        type: string
      profile_id:
        type: integer
      social_media_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para gerenciamento de perfis e currículos com IA.
  title: Resume GenAI API
  version: "1.0"
paths:
...
  /experiences/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Experience'
            type: array
      summary: Lista todas as experiências
      tags:
      - experiences
    post:
      consumes:
      - application/json
      parameters:
      - description: Experiência a ser criada
        in: body
        name: experience
        required: true
        schema:
          $ref: '#/definitions/model.Experience'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Experience'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria uma nova experiência
      tags:
      - experiences
  /experiences/{id}:
    delete:
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove uma experiência
      tags:
      - experiences
    get:
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Experience'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca experiência por ID
      tags:
      - experiences
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da experiência
        in: path
        name: id
        required: true
        type: string
      - description: Experiência atualizada
        in: body
        name: experience
        required: true
        schema:
          $ref: '#/definitions/model.Experience'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Experience'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza uma experiência
      tags:
      - experiences
  /languages/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Language'
            type: array
      summary: Lista todos os idiomas
      tags:
      - languages
    post:
      consumes:
      - application/json
      parameters:
      - description: Idioma a ser criado
        in: body
        name: language
        required: true
        schema:
          $ref: '#/definitions/model.Language'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Language'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria um novo idioma
      tags:
      - languages
  /languages/{id}:
    delete:
      parameters:
      - description: ID do idioma
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove um idioma
      tags:
      - languages
    get:
      parameters:
      - description: ID do idioma
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Language'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca idioma por ID
      tags:
      - languages
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do idioma
        in: path
        name: id
        required: true
        type: string
      - description: Idioma atualizado
        in: body
        name: language
        required: true
        schema:
          $ref: '#/definitions/model.Language'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Language'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza um idioma
      tags:
      - languages
  /profiles/:
    get:
      description: Retorna todos os perfis cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ProfileDTO'
            type: array
      summary: Lista todos os perfis
      tags:
      - profiles
    post:
      consumes:
      - application/json
      description: Cria um novo perfil com os dados informados
      parameters:
      - description: Perfil a ser criado
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/model.ProfileDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ProfileDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria um novo perfil
      tags:
      - profiles
  /profiles/{id}:
    delete:
      description: Remove um perfil pelo ID
      parameters:
      - description: ID do perfil
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove um perfil
      tags:
      - profiles
    get:
      description: Retorna um perfil pelo ID
      parameters:
      - description: ID do perfil
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProfileDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca perfil por ID
      tags:
      - profiles
    put:
      consumes:
      - application/json
      description: Atualiza um perfil existente pelo ID
      parameters:
      - description: ID do perfil
        in: path
        name: id
        required: true
        type: integer
      - description: Perfil atualizado
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/model.ProfileDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProfileDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza um perfil
      tags:
      - profiles
  /profiles/generate:
    post:
      consumes:
      - application/json
      description: Gera um novo perfil usando LLMs com base na descrição da vaga
      parameters:
      - description: Descrição da vaga
        in: body
        name: roleDescription
        required: true
        schema:
          $ref: '#/definitions/dto.RoleDescription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ProfileDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Gera um novo perfil baseado na descrição de vaga
      tags:
      - profiles
...
swagger: "2.0"